test(Storybook): Moved .stories files to test folder
