+++ .storybook/config.js
@@ -8,7 +8,7 @@ const BRConfigs = require('../bluerain');
 addDecorator(BlueRainDecorator(BRConfigs, BR));
 
 // automatically import all files ending in *.stories.js
-const req = require.context('../src', true, /.stories.tsx$/);
+const req = require.context('../test/stories', true, /.stories.tsx$/);
 function loadStories() {
 	req.keys().forEach((filename) => req(filename));
 }
