+++ package.json
@@ -95,7 +95,7 @@
   },
 	"size-limit": [
 		{
-			"limit": "5.1 KB",
+			"limit": "4.8 KB",
 			"webpack": false,
 			"path": "dist/*.js"
 		}
