+++ test/stories/Provider.stories.tsx
@@ -1,4 +1,4 @@
-import { BlueRain, BlueRainConsumer } from './index';
+import { BlueRain, BlueRainConsumer } from '../../src';
 import React from 'react';
 import storiesOf from '@blueeast/bluerain-storybook-addon';
