+++ test/stories/components/Icon.stories.tsx
@@ -1,5 +1,5 @@
-import { withBlueRain } from '../../index';
-import Icon from '../Icon/Icon';
+import { withBlueRain } from '../../../src';
+import Icon from '../../../src/components/Icon/Icon';
 import React from 'react';
 import storiesOf from '@blueeast/bluerain-storybook-addon';
